:recycle: type, constant 분리
