+++ components/PostPreview/PostPreview.stories.tsx
@@ -1,8 +1,7 @@
 import { useCallback } from 'react';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 import PostPreview from './PostPreview';
-
-const LOREM_IPSUM = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.';
+import { LOREM_IPSUM } from '../../utils/constants';
 
 export default {
   component: PostPreview,
