--- pages/api/hello.ts
@@ -1,6 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next';
-// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
-
-export default function handler(req: NextApiRequest, res: NextApiResponse) {
-  res.status(200).json({ name: 'John Doe' });
-}
