+++ pages/list/[page].tsx
@@ -2,7 +2,8 @@
 import { useRouter } from 'next/router';
 import Pagination from '../../components/Pagination/Pagination';
 import PostPreview from '../../components/PostPreview/PostPreview';
-import { getPostCount, getPosts, PostType } from '../../data/posts';
+import { getPosts, getPostCount } from '../../utils/api/posts';
+import { PostType } from '../../types/post';
 
 const POST_PER_LIST = 5;
 const LIST_PER_PAGE = 10;
